requirements.txt
requests
pandas
numpy
chainlit
tensorflow
python-dotenv  
transformers
fastapi 
uvicorn 
docker
pytest 
kaggle 